+++ CHANGELOG.md
@@ -5,6 +5,7 @@ We use [semantic versioning](http://semver.org/):
 - PATCH version when you make backwards compatible bug fixes.
 
 # Next Release
+- [fix] dashboard widget: test gap badge showed wrong percentage
 - [fix] work item integration: findings and test gap badges linked to list of issues instead of issue details in Teamscale
 
 # 1.5.5